requirements.txt
fastapi
uvicorn
httpx
dnspython